# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptOut": {
        "properties": {
          "activity_id": {
            "format": "uuid",
            "title": "Activity Id",
            "type": "string"
          },
          "content_version_id": {
            "format": "uuid",
            "title": "Content Version Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SavedItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "max_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Score"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          },
          "percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percent"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          }
        },
        "required": [
          "id",
          "activity_id",
          "content_version_id",
          "status",
          "started_at",
          "submitted_at",
          "score",
          "max_score",
          "percent",
          "passed"
        ],
        "title": "AttemptOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ItemGradeOut": {
        "properties": {
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "explanation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "item_key": {
            "title": "Item Key",
            "type": "string"
          },
          "max_score": {
            "title": "Max Score",
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "item_key",
          "correct",
          "score",
          "max_score",
          "explanation"
        ],
        "title": "ItemGradeOut",
        "type": "object"
      },
      "ItemIn": {
        "properties": {
          "item_key": {
            "maxLength": 60,
            "title": "Item Key",
            "type": "string"
          },
          "response": {
            "additionalProperties": true,
            "title": "Response",
            "type": "object"
          }
        },
        "required": [
          "item_key",
          "response"
        ],
        "title": "ItemIn",
        "type": "object"
      },
...
      "SavedItemOut": {
        "properties": {
          "correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct"
          },
          "item_key": {
            "title": "Item Key",
            "type": "string"
          },
          "max_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Score"
          },
          "response": {
            "additionalProperties": true,
            "title": "Response",
            "type": "object"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "item_key",
          "response",
          "correct",
          "score",
          "max_score"
        ],
        "title": "SavedItemOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RTApps API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/activities/{activity_id}/attempts": {
      "post": {
        "operationId": "start_attempt_api_v1_activities__activity_id__attempts_post",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Attempt",
        "tags": [
          "attempts"
        ]
      }
    },
...
    "/api/v1/attempts/{attempt_id}/items": {
      "post": {
        "operationId": "grade_item_api_v1_attempts__attempt_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemGradeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grade Item",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/submit": {
      "post": {
        "operationId": "submit_attempt_api_v1_attempts__attempt_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Attempt",
        "tags": [
          "attempts"
        ]
      }
    },
...
  }
}